Added required authentication for endpoints

+++ src/services/membership_service.py
@@ -1,6 +1,8 @@
 from ..repositories.membership_repository import membership_repository
 from ..schemas.membership_schemas import MembershipCreateWithPeriod, MembershipGet
+from ..auth.models import UserModel
 from .base_service import BaseService
+
 
 class MembershipService(BaseService):
     def create(self, data: MembershipCreateWithPeriod) -> MembershipGet:
@@ -21,8 +23,12 @@ class MembershipService(BaseService):
     def get_by_id(self, pk: int) -> MembershipGet:
         return MembershipGet.model_validate(membership_repository.get_single(id=pk))
 
-    def delete_by_id(self, pk: int) -> None:
-        membership_repository.delete(id=pk)
+    def delete_by_id(self, pk: int, user: UserModel) -> bool:
+        for membership in user.memberships:
+            if membership.id == pk:
+                membership_repository.delete(id=pk)
+                return True
+        return False
 
 
 membership_service = MembershipService()
